big bug fix: changed opt_selected_probs_for_rnd input to float tensor instead of long tensor. This was converting everything to 0

+++ train/train_D.py
@@ -254,7 +254,7 @@ def train(epoch):
             wrong_ans_input = torch.LongTensor(wrong_ans.size()).cuda()
             wrong_ans_input.copy_(wrong_ans)
 
-            opt_selected_probs_for_rnd_input = torch.LongTensor(opt_selected_probs_for_rnd.size()).cuda()
+            opt_selected_probs_for_rnd_input = torch.FloatTensor(opt_selected_probs_for_rnd.size()).cuda()
             opt_selected_probs_for_rnd_input.copy_(opt_selected_probs_for_rnd)
 
             # # sample in-batch negative index
